+++ upload/app/code/community/Rees46/Personalization/Helper/Admin.php
@@ -12,6 +12,6 @@ class Rees46_Personalization_Helper_Admin extends Mage_Core_Helper_Abstract
 	 */
 	public function isActionAllowed($action)
 	{
-		return Mage::getSingleton('admin/session')->isAllowed('personalization/manage/' . $action);
+		return Mage::getSingleton('admin/session')->isAllowed('rees46/manage/' . $action);
 	}
 }